Use buildroot to build a toolchain
This is a bit of a strange use of buildroot, since we're really only
interested in the host tools here. But the goal of this is to get the
toolchain built separately from everything else so we can then reuse
it as we build our various image variants.

Normally one would use something like crosstool-ng for this, but since
we're already going to the trouble of downloading buildroot we'll just
use it for this task.