Create "system" and "admin" users
For now, while we're still getting started, these have hard-coded
passwords that are both "temp" so we can log in to test things.

Later the "system" user will get an authorized_keys file containing the
key provided in EC2 metadata and the "admin" user will integrate with
vault's one-time-password SSH setup.

The "admin" user is a sudoer, and is intended to be used by trusted
human sysadmins to get in and debug/fix when things go awry.
The "system" user currently has no special access, but later it might
get some very specific access to be used for automated provisioning-ish
actions.