Produce separate disk images for Xen and QEMU
Amazon EC2 is based on Xen and so it wants Xen-style virtualized IO
devices, but QEMU wants virtio-style devices. This only really manifests
in our bootloader configuration, where we need to use the appropriate
device names for the console and root filesystem.

So we'll produce one root filesystem, but then produce separate boot
filesystems for each platform and then assemble them into two separate
full-disk images so that we can do local testing in QEMU while and yet
still upload results to EC2 for real-world testing and -- eventually --
actual deployment.