Put a banner on the local console after boot
This is mainly just a little easter egg to be seen on systems that allow
the user to take a screenshot of the virtual system's local display.
Later on we'll hopefully put some useful summary information here so that
it can be used for debugging systems that have become otherwise
inaccessible.