Script to set up a virtual test network for QEMU instances
Once we have multiple images in play we'll want to start up more than one
of them in a single network and test how they interact. The test-network
setup script makes this possible by creating a bridge to connect the
qemu TAP devices to, and running dnsmasq on that bridge's interface so
that the machines can get DHCP leases when they boot.

The launch-emulator script is then changed to use the "tap" network
backend with our scripts to join the bridge. As of this change, it now
requires that "test-network start" is run before launching any VMs.

The machines in the private network get access to the internet via NAT
on the host, and are not themselves directly internet-accessible.